requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-login>=0.6
openai>=1.0
python-dotenv>=1.0
